FP6: Reorder initial qcril db to avoid insertion before creation
The ODM added insert statements before the table is created which leads
to build time errors for us. Since we haven't shipped builds yet, we
can fix the initial migration instead of upreving the db and adding a
new migration. This avoids diverging too far from stock and it seems
unlikely that they will notice/fix this.

Change-Id: I5e79b07080a239e411be1acd2e4aa0cfe5d37f9b