FP6: Switch screen density to 440 dpi
Stock sets it to xxxhdpi which is incorrect, this should be a number
and not a density bucket. The result is that everything in the UI
shows larger than it should, both in recovery and userspace. Set it
to 440 dpi since our screen size is similar to FP4 to improve the UI.

Change-Id: I289b21290c5247e2609cc2cb065998a68da179a8